:truck: Restructure meta module
