+++ tests/introspection/test_expression_processing.py
@@ -2,7 +2,8 @@ from textwrap import dedent
 
 import pytest
 
-from json_logic.meta import JSONLogicExpression, Operation, Var
+from json_logic.meta import JSONLogicExpression, Operation
+from json_logic.meta.operations import Var
 from json_logic.typing import Primitive
